// src/shared/api/model/currency.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 1.0.0
 */

export type Currency = typeof Currency[keyof typeof Currency];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const Currency = {
  uzs: 'uzs',
  usd: 'usd',
} as const;