// src/shared/api/model/deadlineType.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 1.0.0
 */

export type DeadlineType = typeof DeadlineType[keyof typeof DeadlineType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DeadlineType = {
  hour: 'hour',
  day: 'day',
  month: 'month',
  year: 'year',
} as const;