// src/shared/api/model/orderInput.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 1.0.0
 */
import type { Currency } from './currency';
import type { DeadlineType } from './deadlineType';

export interface OrderInput {
  author: string;
  /** */
  categoryId: number;
  currency: Currency;
  /** */
  deadline: number;
  deadlineType: DeadlineType;
  price: number;
  skillsIds: number[];
  technicalTask: string;
  title: string;
}